add IOCrypto, UsableAlgorithm RSAEncryption
Rewritten some parts of the AESEncryption class to split reading and writing keys and IVs into separate classes. Also added RSAEncryption class, which generates a KeyPair of a public and private RSA key, and encypts and decrypts symmetric keys and hash files for work in a hybrid crypto system. Work still underway on that class