Add basic gui with starter css and fxml
Add a gui with no functionality

Added a basic gui with no functionality or style. All pages are linked and can be opened by pressing on images or buttons. The interactions between pages will be created later by using the MVC design pattern.

The css, fxml and images are added in the resources folder of the app and will be used to create a responsive design.