Complete decrypt page
Completed the decrypt page

Also fixed a bug where the wrong keys were being used for encrypting and decrypting the hash. The public was used for encrypting and private for decrypting now it is the other way around.

The only thing that can be fixed at a later date is creating an extra button that opens the folder containing only the necessary files for decrypting and not every file. Now the program looks for these files in the folder where the user wants their decrypted message saved. This can create errors but for now they are being handled by giving extra information in the textArea.