Add fully working encryption and decryption functionallity
Completed a working encryption/decryption GUI with the following working functionalities:

A user can create private and public keys.
A user can encrypt a message that he or she can write in a textarea or select a text file with a message in it.
A user can decrypt a message and select where he or she wants to save it.

A window has been added for steganography which will be used later to encrypt messages in an image that the user can choose.