Complete project
Completed the appdev project

Added gui for saving encrypted a message into an image.
Added gui for getting a message out of an image and decrypting this message.
Added folders that are created on startup that simulate the internet and receiver/sender.
Updated labels so the user gets extra information and knows what to do in every window.

Everything works except when a user wants to send more then one message through an image in one session.

Recreation of bug:
1) user creates new keys for sender/receiver
2) user encrypts a message and saves it into an image
3) user decrypts this message and everything works
4) user encrypts a new message and saves it into an image
5) user cannot decrypt the message
6) user restarts program
7) user can decrypt the message